OmniForgeCodex/requirements.txt
# Core dependencies
PySide6>=6.0.0
reportlab>=4.0.0  # For PDF generation
schedule>=1.0.0   # For task scheduling
croniter>=1.0.0   # For cron-style scheduling
psutil>=5.9.0     # For system monitoring
watchdog>=3.0.0   # For file system monitoring
jsonschema>=4.0.0 # For JSON validation
pyyaml>=6.0.0     # For YAML support
requests>=2.0.0   # For HTTP requests
aiohttp>=3.0.0   # For async HTTP requests
cryptography>=41.0.0  # For encryption/decryption
numpy>=1.21.0     # For numerical operations
pandas>=1.3.0     # For data analysis
PyJWT>=2.8.0     # For JWT token handling 
pandas>=1.3.0     # For data analysis 